fix: 박스 조회 및 검색 시 N+1 쿼리 문제 수정 및 최적화

+++ server/src/main/java/teamdropin/server/domain/box/dto/box/UpdateBoxRequestDto.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import teamdropin.server.domain.box.dto.boxTag.UpdateBoxTagRequestDto;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
@@ -17,27 +18,19 @@ import java.util.List;
 public class UpdateBoxRequestDto {
 
     private Long id;
-
     @NotBlank
     private String name;
-
     @NotBlank
     private String location;
-
     private String phoneNumber;
-
     @NotNull
     private Long cost;
-
     @NotNull
     private Long area;
-
     private boolean barbellDrop;
     private String url;
     private String detail;
-
     private List<String> tagList;
-
-    private HashMap<Integer,String> imageInfo;
+    private HashMap<Long,String> imageInfo;
 
 }
